fixed bugs, hidden jwt_secret and added owners to tracks/albums

+++ src/track/track.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException, Logger } from "@nestjs/common";
+import { Injectable, NotFoundException, Logger, InternalServerErrorException, ForbiddenException } from "@nestjs/common";
 import { InjectModel } from "@nestjs/mongoose";
 import { Track, TrackDocument } from "./schemas/track.schema";
 import { Model, ObjectId, Types } from "mongoose";
@@ -11,19 +11,15 @@ import { Album, AlbumDocument } from "src/album/schemas/album.schema";
 
 @Injectable()
 export class TrackService {
-    private readonly logger = new Logger(TrackService.name);
-
     constructor(@InjectModel(Track.name) private trackModel: Model<TrackDocument>,
                 @InjectModel(Comment.name) private commentModel: Model<CommentDocument>,
                 @InjectModel(Album.name) private albumModel: Model<AlbumDocument>,
                 private fileService: FileService) {}
 
-    async create(dto: CreateTrackDto, audio, cover): Promise<Track> {
+    async create(dto: CreateTrackDto, userId: string, artistName: string, audio, cover): Promise<Track> {
         const auidoPath = this.fileService.createFile(FileType.AUDIO, audio);
         const coverPath = this.fileService.createFile(FileType.IMAGE, cover);
-        const track = await this.trackModel.create({...dto, listens: 0, audio: auidoPath, cover: coverPath});
-
-        this.logger.log("Created track")
+        const track = await this.trackModel.create({...dto, artist: artistName, owner: userId, listens: 0, audio: auidoPath, cover: coverPath});
 
         return track;
     }
@@ -32,24 +28,14 @@ export class TrackService {
         const track = await this.trackModel.findById(id).populate("comments");
 
         if(!track) {
-            this.logger.error("failed to get track");
             throw new NotFoundException("track not found");
         }
-
-        this.logger.log("getted track");
 
         return track;
     }
 
     async getAll(limit = 10, offset = 0): Promise<Track[]> {
         const tracks = await this.trackModel.find().skip(offset).limit(limit);
-
-        if(!tracks) {
-            this.logger.error("failed to get all tracks")
-            throw new NotFoundException("failed to get all tracks");
-        }
-
-        this.logger.log("getted all tracks");
 
         return tracks;
     }  
@@ -64,94 +50,98 @@ export class TrackService {
 
 
         if (tracks.length === 0) {
-            this.logger.error("track not found")
             return [];
         }
-
-        this.logger.log("finded tracks with")
 
         return tracks;
     }
 
-    async delete(id: Types.ObjectId): Promise<Types.ObjectId> {
-        const track = await this.trackModel.findOneAndDelete({_id: id});
+    async delete(id: Types.ObjectId, ownerId: string): Promise<Types.ObjectId> {
+        const track = await this.trackModel.findOne({_id: id});
 
         if(!track) {
-            this.logger.error("track not found")
             throw new NotFoundException("track is not exists");
         }
 
-        this.logger.log("deleted track");
+        if(track.owner.toString() !== ownerId) {
+            throw new ForbiddenException("don't have the rights to do this")
+        }
+
+        await track.deleteOne({_id: id});
 
         return track._id;
     }  
 
     async addComment(dto: CreateCommentDto): Promise<Comment> {
-        const track = await this.trackModel.findById(dto.trackId);
+        const track = await this.trackModel.findOne({_id: dto.trackId});
+
+        console.log(track)
+
+        if(!track) {
+            throw new NotFoundException("track is not exists");
+        }
+
         const comment = await this.commentModel.create({...dto})
 
         if(!comment) {
-            this.logger.error("failed to create comment")
-            throw new NotFoundException("fail to create comment");
-        }
-
-        if(!track) {
-            this.logger.error("track is not exists")
-            throw new NotFoundException("track is not exists");
+            throw new InternalServerErrorException("fail to create comment");
         }
 
         track.comments.push(comment._id);
         await track.save();
 
-        this.logger.log("added comment");
-
         return comment;
     }
 
-    async deleteComment(id: Types.ObjectId): Promise<Types.ObjectId> {
-        const comment = await this.commentModel.findOneAndDelete({_id: id});
+    async deleteComment(id: Types.ObjectId, ownerId: string): Promise<Types.ObjectId> {
+        const comment = await this.commentModel.findOne({_id: id});
 
         if(!comment) {
             throw new NotFoundException("comment is not exists")
         }
 
-        await this.trackModel.findByIdAndUpdate(comment.track, {$pull: {comments: id}});
+        if(comment.owner.toString() !== ownerId) {
+            throw new ForbiddenException("don't have the rights to do this")
+        }
 
-        this.logger.log("deleted comment");
+        await this.trackModel.findByIdAndUpdate(comment.track, {$pull: {comments: id}});
 
         comment.deleteOne();
 
         return comment._id;
     }
 
-    async addTrackToAlbum(id: Types.ObjectId, dto: AddTrackDto, audio) {
+    async addTrackToAlbum(id: Types.ObjectId, dto: AddTrackDto, ownerId: string, audio) {
         const album = await this.albumModel.findById(id);
-
-        this.logger.log(id);   
-
-        const auidoPath = this.fileService.createFile(FileType.AUDIO, audio);
-        const track = await this.trackModel.create({...dto, listens: 0, audio: auidoPath, album: id});
 
         if(!album) {
             throw new NotFoundException("failed to find album by id");
         }
 
+        if(album.owner.toString() !== ownerId) {
+            throw new ForbiddenException("don't have the rights to do this")
+        }
+
+        const auidoPath = this.fileService.createFile(FileType.AUDIO, audio);
+        const track = await this.trackModel.create({...dto, listens: 0, audio: auidoPath, album: id});
+
+        if(!track) {
+            throw new InternalServerErrorException("failed to create track by album");
+        }
+
         album.tracks.push(track._id);
         await album.save();
-
-        this.logger.log("track added to album")
 
         return track;
     }
 
     async listen(id: Types.ObjectId) {
-        const track = await this.trackModel.findById({_id: id});
+    const track = await this.trackModel.findByIdAndUpdate(id, { $inc: { listens: 1 }}, { new: true} );
 
-        if(!track) {
-            throw new NotFoundException("track not found")
-        }
-
-        track.listens += 1;
-        track.save();
+    if (!track) {
+        throw new NotFoundException("track not found");
     }
+
+    return track;
+}
 }
